+++ src/app/user/layout.tsx
@@ -2,7 +2,7 @@ import type {Metadata} from "next";
 import "@/app/globals.css";
 import Header from "@/components/Header";
 import HeaderV2 from "@/components/HeaderV2";
-import AuthAside from "@/app/auth/components/AuthAside";
+import AuthAside from "@/components/auth/AuthAside";
 import Footer from "@/components/Footer";
 
 export const metadata: Metadata = {
