+++ src/components/auth/OtherAuth.tsx
@@ -1,11 +1,11 @@
 import Link from "next/link";
 
-type OtherAuthProps = {
+interface OtherAuthProps {
     value: string,
     link: string
 }
 
-const OtherAuth: React.FC<OtherAuthProps> = ({value, link}) =>{
+export const OtherAuth: React.FC<OtherAuthProps> = ({value, link}) =>{
     return(
         <Link href={`${link}`} className='h-[74px] w-[205px] p-[16px] border-black border-2 flex items-center text-[16px] cursor-pointer'>
             <section className="w-2/3">
@@ -18,4 +18,3 @@ const OtherAuth: React.FC<OtherAuthProps> = ({value, link}) =>{
     );
 };
 
-export default OtherAuth;
